+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [UIMARCAUTH-409](https://issues.folio.org/browse/UIMARCAUTH-409) React v19: refactor away from default props for functional components.
 - [UIMARCAUTH-461](https://issues.folio.org/browse/UIMARCAUTH-461) Upgrade `browse` to `2.0`.
 - [UIMARCAUTH-465](https://issues.folio.org/browse/UIMARCAUTH-465) Add `MarcVersionHistory` to display all versions in the fourth pane.
+- [UIMARCAUTH-463](https://issues.folio.org/browse/UIMARCAUTH-463) AuthorityView - add tenantId to useAuditSettings.
 
 ## [6.0.1](https://github.com/folio-org/ui-marc-authorities/tree/v6.0.1) (2024-12-13)